+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+- Error when vacation days is object instead of array
+
 ## [1.1.1] - 2022-06-15
 
 ### Added